separate gpu and nvml
